pyproject.toml
[project]
name = "livekit-langgraph-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "livekit-agents[deepgram,langchain]~=1.4",
    "livekit-api>=1.1.0",
    "livekit-plugins-openai>=1.5.1",
    "livekit-plugins-silero>=1.5.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
